made break_website_into_commits.bash a good working script for mac bash version 3 bigsur tested

+++ break_website_into_commits.bash
@@ -224,7 +224,7 @@ function _breakable_copy_directory(){
 	local _target="${_to}/${_new_folder_name}"
   (( DEBUG )) && echo "_target:${_target}"
   enforce_variable_with_value _target "${_target}"
-	local _date="$(cut -d- -f4- <<<"${_new_folder_name}")"
+	local _date="$(tail -1<<<"$(sed 's/-/\n/g'<<<"${_new_folder_name}")")"
   (( DEBUG )) && echo "_date:${_date}"
   enforce_variable_with_value _date "${_date}"
   Checking "_date :${_date}:"
@@ -234,8 +234,13 @@ function _breakable_copy_directory(){
   enforce_variable_with_value _log_path "${_log_path}"
 	cd "${_from}"
 	local _checked_out="$(/usr/bin/git checkout "${_commit}" 2>&1)"
+	_err=$?
+	wait
+	[ $_err -gt 0 ] && log "${_log_path}" "_FAILED_" "git checkout ${_commit}"
 	/usr/bin/git log -1 --oneline > COMMIT
 	_err=$?
+	wait
+	[ $_err -gt 0 ] && log "${_log_path}" "_FAILED_" "git  log -1 --oneline "
 	local _last_commit="$(<COMMIT)"
 	enforce_variable_with_value _last_commit "${_last_commit}"
 	local _last_lower="$(to_lowercase "${_last_commit}")"
@@ -246,6 +251,9 @@ function _breakable_copy_directory(){
 		[[ "${_last_lower}" == *"dependabot"* ]] || \
 		[[ "${_last_lower}" == *"cleanup"* ]] || \
 		[[ "${_last_lower}" == *"bump"* ]] || \
+		( [[ "${_last_lower}" == *"update"* ]] && [[ "${_last_lower}" == *"depency"* ]] ) || \
+ 		( [[ "${_last_lower}" == *"update"* ]] && [[ "${_last_lower}" == *"dep"* ]] ) || \
+ 		( [[ "${_last_lower}" == *"update"* ]] && [[ "${_last_lower}" == *"bundle"* ]] ) || \
 		( [[ "${_last_lower}" == *"deps"* ]] && [[ "${_last_lower}" == *"bump"* ]] ) || \
 		( [[ "${_last_lower}" == *"deps"* ]] && [[ "${_last_lower}" == *"build"* ]] )  \
    ; then
@@ -261,6 +269,8 @@ function _breakable_copy_directory(){
 		sed 's/merge pull request #....//g'  | \
 		sed 's/ from//g' | \
 		sed 's/ fix//g' | \
+		sed 's/ fixed//g' | \
+		sed 's/ bug//g' | \
 		sed 's/ add//g' | \
 		sed 's/ branch//g' | \
 		sed 's/branch//g' | \
@@ -320,10 +330,24 @@ function _breakable_copy_directory(){
 	[  $_err -gt 0 ] && touch "${_new_folder_name}._FAILED_.RM" && return 1
 	[ $_err -gt 0 ] && log "${_log_path}" "_FAILED_" "RM" && return 1
 	Message $BASHLINENO $(pwd)
-	Message linking ln -s ../../node_modules node_modules
-	ln -s ../../node_modules node_modules
-	Message linking ln -s ../../vendor vendor
-	ln -s ../../vendor vendor
+	if it_exists_with_spaces ../../node_modules ; then
+	{
+		Message linking ln -s ../../node_modules node_modules
+		ln -s ../../node_modules node_modules
+	}
+  fi
+	if it_exists_with_spaces ../../vendor ; then
+	{
+		Message linking ln -s ../../vendor vendor
+		ln -s ../../vendor vendor
+	}
+  fi
+	if it_exists_with_spaces ../../.idea ; then
+	{
+		Message linking ln -s ../../.idea .idea
+		ln -s ../../.idea .idea
+	}
+  fi
 	# if ("${_packager}" install 2>&1) ; then
 	# {
 	# 	_err=$?
@@ -343,6 +367,12 @@ function _breakable_copy_directory(){
   #  [ $_err -gt 0 ] && log "${_log_path}" "_FAILED_" "TEST"
   # Patchy thingy here
   # second attempt to get the of the folder right when the folder name defaults to :123---- like this
+  local _again_folder_name7=$(only_7_dashed_words<<<"${_again_folder_name}")
+  if [[ -n "${_again_folder_name7}" ]] ; then
+  {
+		_again_folder_name="${_again_folder_name7}"
+	}
+  fi
   if it_exists_with_spaces "${_to}/${_commit_number}_${_again_folder_name}_${_date}" ; then
 	{
 		Message rm -rf "${_to}/${_commit_number}_${_again_folder_name}_${_date}"
@@ -364,8 +394,24 @@ function _breakable_copy_directory(){
 # 1. All commits
 # 2. Loop
 # 3. cp -R
-function _one_liner(){                    # grep tab char       # delete_empty_lines
-	echo "$(cat "${1}/breakage_sizes.log" | grep -P '\t'"${2}_")" |  sed '/^\s*$/d' >>"${1}/breakage_sizes_sorted.log"
+function _one_liner(){
+# 	tests for mac
+# echo "100M	99_9d448276-update-bundler-to-depency-error-with_20180313T085412Z
+# 100M	9_9d448276-update-bundler-to-depency-error-with_more-view-20180212T230031Z" | pcregrep '\t'"9_"  |  sed '/^\s*$/d'
+# echo -n "100M   99_9d448276-update-bundler-to-depency-error-with_20180313T085412Z
+# 100M    9_9d448276-update-bundler-to-depency-error-with_more-view-20180212T230031Z" | pcregrep '\t'"99_" | sed '/^\s*$/d'
+	if [[ "$(uname)" == "Darwin" ]] ; then
+	{
+	  # echo "# mac grep"                      # grep tab char       # delete_empty_lines
+		echo "$(cat "${1}/breakage_sizes.log" | pcregrep '\t'"${2}")" |  sed '/^\s*$/d' >>"${1}/breakage_sizes_sorted.log"
+	}
+	else
+	{
+	  # echo "# windows /linux grep"           # grep tab char       # delete_empty_lines
+		echo "$(cat "${1}/breakage_sizes.log" | grep -P '\t'"${2}")" |  sed '/^\s*$/d' >>"${1}/breakage_sizes_sorted.log"
+	}
+	fi
+  #
 } # end _one_liner
 
 function _break_all_commits(){
@@ -455,6 +501,9 @@ function _break_all_commits(){
   			[[ "${_last_lower}" == *"build(deps)"* ]] ||
   			[[ "${_last_lower}" == *"bdependabot"* ]] ||
   			[[ "${_last_lower}" == *"bump"* ]] ||
+    		( [[ "${_last_lower}" == *"update"* ]] && [[ "${_last_lower}" == *"depency"* ]] ) || \
+    		( [[ "${_last_lower}" == *"update"* ]] && [[ "${_last_lower}" == *"dep"* ]] ) || \
+    		( [[ "${_last_lower}" == *"update"* ]] && [[ "${_last_lower}" == *"bundle"* ]] ) || \
   			( [[ "${_last_lower}" == *"deps"* ]] && [[ "${_last_lower}" == *"bump"* ]] ) ||
   			( [[ "${_last_lower}" == *"deps"* ]] && [[ "${_last_lower}" == *"build"* ]] )
   		) ; then
@@ -484,19 +533,20 @@ function _break_all_commits(){
 	while read -r _commit ; do
 	{
 		[[ -z "${_commit}" ]] && continue  # skip empties
-		_one_liner "${TARGETFOLDER}" "${_current_commit}"
+		_one_liner "${TARGETFOLDER}" "${_current_commit}_"
 		_current_commit=$(( _current_commit - 1 ))
   }
 	done <<< "${_commits}"
   # done
-	echo "$(cat "${TARGETFOLDER}/breakage_sizes.log" | grep -P '\t'"total")" |  sed '/^\s*$/d' >>"${TARGETFOLDER}/breakage_sizes_sorted.log"
+  _one_liner "${TARGETFOLDER}" "total"
+	# echo "$(cat "${TARGETFOLDER}/breakage_sizes.log" | grep -P '\t'"total")" |  sed '/^\s*$/d' >>"${TARGETFOLDER}/breakage_sizes_sorted.log"
 		return 0
 } # end break_all_commits
 # check_required_packages pnpm "
 #   ava
 
 # "
-if [ -z ${1} ] ; then
+if [ -z ${1} ] || [ -z ${2} ]  || [ -z ${3} ] ; then
 {
 echo "
 
@@ -510,8 +560,7 @@ $0 /home/user/_/work/client/rnd/project /home/zeus/_/work/client/rnd/project_bre
 }
 else
 {
-	# "${*}"
-	_break_all_commits ${*}
+	_break_all_commits "${1}" "${2}" "${3}"
 }
 fi
 echo "🥦"
